Make draw work with an NxN size board if N < 100.
Next steps: Make draw work with an NxN size board for any N.
The challenge would be in making the initial space for a line variable
with magnitude of N. Right now it's hardcoded for N < 100, or
magnitude(n) < 3. The spacing at the start of a line should be
magnitude(n) + 1, this would leave a single space before the row
dividers and 2 spaces before the player mark.